HW2 test cases correct

+++ src/homework/01_variables/variables.cpp
@@ -1,4 +1,5 @@
 #include "variables.h"
+#include "auto.h"
 
 //example
 int add_numbers(int num1, int num2)
@@ -9,7 +10,7 @@ int add_numbers(int num1, int num2)
 //write code for multiply_numbers function here
 //function returns product of num1 and num2
 
-int get_sales_tax(double meal_amount)
+double get_sales_tax(double meal_amount)
 {
 	const double tax_rate = 0.0675;
 
@@ -17,7 +18,7 @@ int get_sales_tax(double meal_amount)
 }
 
 
-int get_tip_amount(double meal_amount, double tip_rate)
+double get_tip_amount(double meal_amount, double tip_rate)
 {
 
 	return meal_amount * tip_rate;
